requirements.txt
# --- RL / Simulation ---
gymnasium>=1.0.0
pygame>=2.5.0
numpy
stable-baselines3

# --- Agent framework ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic
langchain-openai
langchain-google-genai

# --- Backend ---
fastapi
uvicorn[standard]
websockets
python-dotenv
python-multipart

# --- Articulate-Anything (URDF generation) ---
-e core/tools/articulate_repo/
hydra-core
omegaconf
trimesh
rich
termcolor
astor
markdown2
google-generativeai
ipython

# --- Utilities ---
Pillow
